fix: refactor processor function
